\name{msa_example}
\docType{data}
\alias{msa_example}
\title{msa_example}
\usage{
msa_example
}
\description{
The multiple structure alignment information, by incorporating it into a matrix, where 1) the number of rows and columns of the SYMMETRIC and NON-NEGATIVE matrix should be the length of amino acid sequence; 2) each element give the number of times that each pair of residuals is contacted in the multiple structure alignment; 3) elements are separated by white space.
}

